+++ Translation/Dumper/XliffDumper.php
@@ -32,9 +32,24 @@ use JMS\TranslationBundle\Model\MessageCatalogue;
  */
 class XliffDumper implements DumperInterface
 {
+    /**
+     * @var string
+     */
     private $sourceLanguage = 'en';
+
+    /**
+     * @var bool
+     */
     private $addDate = true;
+
+    /**
+     * @var bool
+     */
     private $addReference = true;
+
+    /**
+     * @var bool
+     */
     private $addReferencePosition = true;
 
     /**
@@ -42,7 +57,7 @@ class XliffDumper implements DumperInterface
      */
     public function setAddDate($bool)
     {
-        $this->addDate = (Boolean) $bool;
+        $this->addDate = (bool) $bool;
     }
 
     /**
@@ -70,7 +85,8 @@ class XliffDumper implements DumperInterface
     }
 
     /**
-     * @param \JMS\TranslationBundle\Model\MessageCatalogue $domain
+     * @param MessageCatalogue $catalogue
+     * @param MessageCatalogue|string $domain
      * @return string
      */
     public function dump(MessageCatalogue $catalogue, $domain = 'messages')
@@ -159,7 +175,6 @@ class XliffDumper implements DumperInterface
             if ($meaning = $message->getMeaning()) {
                 $unit->setAttribute('extradata', 'Meaning: '.$meaning);
             }
-
         }
 
         return $doc->saveXML();