+++ Changelog.md
@@ -1,9 +1,17 @@
 # Changelog
 
 
-### 1.2.1 to next release
+### 1.2.2 to next release
+
+
+### 1.2.1 to 1.2.2
 
 * Support nikic/php-parser 1.4.x and 2.0.x
+* Set XLF as default output format of the ExtractTranslationCommand
+* Better compatibility with Symfony 3.0
+* Bugfixes with calls to the logger
+* Added more tests for controllers, extension, compiler passes and config
+* Code and doc cleanup 
 
 ### 1.2.0 to 1.2.1